RPC_PyTorch/requirements.txt
psutil==5.9.0
python-dotenv==1.1.0
setuptools==59.6.0
torch==1.12.0
torchvision==0.13.0
numpy==1.23.5